nx: set up nx
nx feels a small step more complicated than turborepo with pretty
similar output.  It _does_ support building only a certain subtree of
packages, but in practice that may not be terribly useful: when an
interior node of the dependency DAG changes, nx allows only that that
interior node and its dependencies to be rebuilt. While this does allow
for slightly faster builds and less work to be performed, it also
critically allows those leaf nodes to not detect breaking changes from
the interior node. :emoji_of_a_thinking_face_emoji: